Split main reducer from item reducer
